% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reminder.R
\name{reminder}
\alias{reminder}
\title{A Heartfelt Reminder}
\usage{
reminder(name = "Jennifer")
}
\arguments{
\item{name}{Defaults to "Jennifer" (probably because she's awesome).}
}
\description{
This function allows you to remind yourself or someone else that they are awesome.
}
\examples{
reminder("Collette")
}
\keyword{cats}

