% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reminder.R
\name{reminder}
\alias{reminder}
\title{A Heartfelt Reminder}
\usage{
reminder(name = "Jennifer", font = "Helvetica")
}
\arguments{
\item{name}{Defaults to "Jennifer" (probably because she's awesome)}

\item{font}{Defaults to "Helvetica"}
}
\description{
This function allows you to remind yourself or someone else that they are awesome.
}
\examples{
reminder(name = "Lisa", font = 'Mistral')
}
\keyword{maybe}
\keyword{miguel}
  
